requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
reportlab==4.0.9
qrcode[pil]==7.4.2
fastapi-mail==1.4.1
pillow==10.2.0
httpx==0.26.0
